REFACTORED: location of pubsub.js
