+++ lib/pubsub.js
