+++ package.json
@@ -2,7 +2,7 @@
   "name": "amqplib-pubsub",
   "version": "0.0.1",
   "description": "Simplified API for Pub/Sub with amqplib (e.g. RabbitMQ)",
-  "main": "pubsub.js",
+  "main": "/lib/pubsub.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
